Fix all the version problems after trying to clone
I had a problem where my proxy would be used for my get requests but
not my post requests. Stil havnt figured out why. Wanted to clean my
cache etc so deleted pakcages-lock and node modules. This was a bad
idea. Took ages to fix it and now we are back to the beginnign of the
frontend authentication.