Complete change user informationn along with alerts
We created a profile page with a form which pulled our user info from
our state and then populated our form with it. The changes of the form
were then posted to our api backend and then our state and local
storage was updated. Gonna have lots to go through...