Complete order screen and fix redux bug
We completed OrderScreen.js and finally found out why we werent
assigningn our shipping address back to our state. It was because we
were trying to access obj.shipping_address instead of
obj.shippingaddress. Don't ask why the fuck its all lowercase one word
because I don't know. It was a problem in the serializer. We were trying
to add the shipping address to be returned with the api call and thats
done through those custom serializer things.