Got paypal working... Kinda
The struggles of PayPal. I don't know why but I can't get paypal to work
correctly. First we created the sandbox accounts and then the apps and
then we setup a script to be added to the dom when on the order screen
and teh we installed react-paypal-buttons-v2 so we could implement the
buttons. But the buttons don't go away after we pay until we refresh and
some of the logic is being weird etc.