Implemented our admin user list
We created a way for only admins to look at a list of users and then
delete users. I had a problem with auto rerendering which turned out to
be problems with updating the state with our useeffect and its
dependencies. I then fixed it by looking at state and making sure we
were pulling the correct data. I **think this is the same problem with
our paypal not updating**