Add image upload ability
We can now upload an image into our frontend and have it be sent back to
django in the back. We had to go to setting and add multiform/form-data
to the setting parser to be allowed to upload an image. We also made it
so that when you upload a new image, you delete the old one. In the
future it's probably doable to link one image to one product and cascade
on delete or something idk.