Add functionality to mark as delivered
We can now mark as delivered on the order page. The paypal buttons are
still being weird but this is nothing I wouldn't be able to fix if this
was truly my own build from scratch. Just cant be arse to trace the
logic right now to source the problemo.